% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acOpt.R
\name{acOpt}
\alias{acOpt}
\title{acOpt function}
\usage{
acOpt(x, n, targetFunc, pherFunc, xAt0 = FALSE, evapRate = 0.05,
  nAnts = 500, maxIter = 300, countThresh = 1000,
  showResult = FALSE, pherPower = 1, dumbAnt = FALSE,
  plateau = 100, returnStats = FALSE)
}
\arguments{
\item{x}{[value]}

\item{n}{[value]}

\item{targetFunc}{[value]}

\item{pherFunc}{[value]}

\item{xAt0}{[value]. Default is FALSE}

\item{evapRate}{[value]. Default is 0.05}

\item{nAnts}{[value]. Default is 500}

\item{maxIter}{[value]. Default is 300}

\item{countThresh}{[value]. Default is 1000}

\item{showResult}{[value]. Default is FALSE}

\item{pherPower}{[value]. Default is 1}

\item{dumbAnt}{[value]. Default is FALSE}

\item{plateau}{[value]. Default is 100}

\item{returnStats}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
