% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACquant.R
\name{ACquant}
\alias{ACquant}
\title{ACquant function}
\usage{
ACquant(pop, GSfit, nSel, nCrosses, use, acTrunc = 1, evapRate = 0.05,
  nAnts = 500, pherPower = 1.5, verbose = FALSE, nProgeny = 1, ...)
}
\arguments{
\item{pop}{[value]}

\item{GSfit}{[value]}

\item{nSel}{[value]}

\item{nCrosses}{[value]}

\item{use}{[value]}

\item{acTrunc}{[value]. Default is 1}

\item{evapRate}{[value]. Default is 0.05}

\item{nAnts}{[value]. Default is 500}

\item{pherPower}{[value]. Default is 1.5}

\item{verbose}{[value]. Default is FALSE}

\item{nProgeny}{[value]. Default is 1}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
