% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argumentChange.R
\name{argumentChange}
\alias{argumentChange}
\title{argumentChange function}
\usage{
argumentChange(defaultArgs, userArgs)
}
\arguments{
\item{defaultArgs}{[value]}

\item{userArgs}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
