% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFit.R
\name{checkFit}
\alias{checkFit}
\title{checkFit function}
\usage{
checkFit(pop)
}
\arguments{
\item{pop}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
