% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dFSel.R
\name{dFSel}
\alias{dFSel}
\title{dFSel function}
\usage{
dFSel(dF, limit = 1, val = "selCrit", parentCols = c("p1", "p2"),
  returnPar = TRUE)
}
\arguments{
\item{dF}{[value]}

\item{limit}{[value]. Default is 1}

\item{val}{[value]. Default is "selCrit"}

\item{parentCols}{[value]. Default is c("p1", "p2")}

\item{returnPar}{[value]. Default is TRUE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
