% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyFunc.R
\name{dummyFunc}
\alias{dummyFunc}
\title{dummyFunc function}
\usage{
dummyFunc(x, retrn)
}
\arguments{
\item{x}{[value]}

\item{retrn}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
