% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estIntensity.R
\name{estIntensity}
\alias{estIntensity}
\title{estIntensity function}
\usage{
estIntensity(VDP, i, nT = nTrial, start = "trial1", end = "variety",
  estFunc = pheno, Gvar = estVg, simParam = NULL)
}
\arguments{
\item{VDP}{[value]}

\item{i}{[value]}

\item{nT}{[value]. Default is nTrial}

\item{start}{[value]. Default is "trial1"}

\item{end}{[value]. Default is "variety"}

\item{estFunc}{[value]. Default is pheno}

\item{Gvar}{[value]. Default is estVg}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
