% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estVg.R
\name{estVg}
\alias{estVg}
\title{estVg function}
\usage{
estVg(pop, GSfit = NULL, GSfunc = RRBLUP)
}
\arguments{
\item{pop}{[value]}

\item{GSfit}{[value]. Default is NULL}

\item{GSfunc}{[value]. Default is RRBLUP}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
