% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDist.R
\name{expDist}
\alias{expDist}
\title{expDist function}
\usage{
expDist(nSel, pop, GSfit, use, quant, returnQuant = TRUE,
  pullGeno = pullSnpGeno, updateEBV = FALSE, weight = 0.5,
  nProgeny = 1, ...)
}
\arguments{
\item{nSel}{[value]}

\item{pop}{[value]}

\item{GSfit}{[value]}

\item{use}{[value]}

\item{quant}{[value]}

\item{returnQuant}{[value]. Default is TRUE}

\item{pullGeno}{[value]. Default is pullSnpGeno}

\item{updateEBV}{[value]. Default is FALSE}

\item{weight}{[value]. Default is 0.5}

\item{nProgeny}{[value]. Default is 1}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
