% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDistPairs.R
\name{expDistPairs}
\alias{expDistPairs}
\title{expDistPairs function}
\usage{
expDistPairs(pop, GSfit, nSel, quant, nCrosses, use, returnQuant = TRUE,
  weightLoci = FALSE, pullGeno = pullSnpGeno, maxCrossPerParent = 0,
  Gvar = estVg, weight = 0.5, nProgeny = 1, verbose = FALSE, ...)
}
\arguments{
\item{pop}{[value]}

\item{GSfit}{[value]}

\item{nSel}{[value]}

\item{quant}{[value]}

\item{nCrosses}{[value]}

\item{use}{[value]}

\item{returnQuant}{[value]. Default is TRUE}

\item{weightLoci}{[value]. Default is FALSE}

\item{pullGeno}{[value]. Default is pullSnpGeno}

\item{maxCrossPerParent}{[value]. Default is 0}

\item{Gvar}{[value]. Default is estVg}

\item{weight}{[value]. Default is 0.5}

\item{nProgeny}{[value]. Default is 1}

\item{verbose}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
