% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCov.R
\name{genCov}
\alias{genCov}
\title{genCov function}
\usage{
genCov(M, u = NULL, absU = TRUE, sumVar = TRUE, scaleD = TRUE,
  inclm = TRUE, p = NULL)
}
\arguments{
\item{M}{[value]}

\item{u}{[value]. Default is NULL}

\item{absU}{[value]. Default is TRUE}

\item{sumVar}{[value]. Default is TRUE}

\item{scaleD}{[value]. Default is TRUE}

\item{inclm}{[value]. Default is TRUE}

\item{p}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
