% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCov2.R
\name{genCov2}
\alias{genCov2}
\title{genCov2 function}
\usage{
genCov2(M, u = NULL, absD = TRUE, sumVar = TRUE, scaleD = TRUE,
  inclm = TRUE, p = NULL)
}
\arguments{
\item{M}{[value]}

\item{u}{[value]. Default is NULL}

\item{absD}{[value]. Default is TRUE}

\item{sumVar}{[value]. Default is TRUE}

\item{scaleD}{[value]. Default is TRUE}

\item{inclm}{[value]. Default is TRUE}

\item{p}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
