% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAcc.R
\name{getAcc}
\alias{getAcc}
\title{getAcc function}
\usage{
getAcc(pop, simParam)
}
\arguments{
\item{pop}{[value]}

\item{simParam}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
