% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAF.R
\name{getAF}
\alias{getAF}
\title{getAF function}
\usage{
getAF(pop, pullGeno = pullSnpGeno, simParam = NULL)
}
\arguments{
\item{pop}{[value]}

\item{pullGeno}{[value]. Default is pullSnpGeno}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
