% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArgs.R
\name{getArgs}
\alias{getArgs}
\title{getArgs function}
\usage{
getArgs(f, ...)
}
\arguments{
\item{f}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
