% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComArgs.R
\name{getComArgs}
\alias{getComArgs}
\title{getComArgs function}
\usage{
getComArgs(defaultArgs = NULL)
}
\arguments{
\item{defaultArgs}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
