% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getF.R
\name{getF}
\alias{getF}
\title{getF function}
\usage{
getF(pop, pullGeno = pullSnpGeno)
}
\arguments{
\item{pop}{[value]}

\item{pullGeno}{[value]. Default is pullSnpGeno}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
