% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geth2.R
\name{geth2}
\alias{geth2}
\title{geth2 function}
\usage{
geth2(pop, GSfit)
}
\arguments{
\item{pop}{[value]}

\item{GSfit}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
