% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopMeanVar.R
\name{getPopMeanVar}
\alias{getPopMeanVar}
\title{getPopMeanVar function}
\usage{
getPopMeanVar(parVal, parCov, Vg)
}
\arguments{
\item{parVal}{[value]}

\item{parCov}{[value]}

\item{Vg}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
