% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopStats.R
\name{getPopStats}
\alias{getPopStats}
\title{getPopStats function}
\usage{
getPopStats(resultL, simParam, meanVariety = TRUE, verbose = FALSE)
}
\arguments{
\item{resultL}{[value]}

\item{meanVariety}{[value]. Default is TRUE}

\item{verbose}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
