% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRRh2.R
\name{getRRh2}
\alias{getRRh2}
\title{getRRh2 function}
\usage{
getRRh2(rrFit)
}
\arguments{
\item{rrFit}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
