% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSel.R
\name{getSel}
\alias{getSel}
\title{getSel function}
\usage{
getSel(selCrit, n, high = TRUE, variable = "selCrit",
  parentCols = c("p1", "p2"), maxP = NULL)
}
\arguments{
\item{selCrit}{[value]}

\item{n}{[value]}

\item{high}{[value]. Default is TRUE}

\item{variable}{[value]. Default is "selCrit"}

\item{parentCols}{[value]. Default is c("p1", "p2")}

\item{maxP}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
