% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSelfVar.R
\name{getSelfVar}
\alias{getSelfVar}
\title{getSelfVar function}
\usage{
getSelfVar(M, u = NULL, fdiff = NULL)
}
\arguments{
\item{M}{[value]}

\item{u}{[value]. Default is NULL}

\item{fdiff}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
