% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTotalIntensity.R
\name{getTotalIntensity}
\alias{getTotalIntensity}
\title{getTotalIntensity function}
\usage{
getTotalIntensity(x)
}
\arguments{
\item{x}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
