% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrueBV.R
\name{getTrueBV}
\alias{getTrueBV}
\title{getTrueBV function}
\usage{
getTrueBV(pop, simParam, trait = 1)
}
\arguments{
\item{pop}{[value]}

\item{simParam}{[value]}

\item{trait}{[value]. Default is 1}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
