% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrueQTLeff.R
\name{getTrueQTLeff}
\alias{getTrueQTLeff}
\title{getTrueQTLeff function}
\usage{
getTrueQTLeff(simParam, trait = 1)
}
\arguments{
\item{simParam}{[value]}

\item{trait}{[value]. Default is 1}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
