% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYrange.R
\name{getYrange}
\alias{getYrange}
\title{getYrange function}
\usage{
getYrange(simR)
}
\arguments{
\item{simR}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
