% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsQuant.R
\name{gsQuant}
\alias{gsQuant}
\title{gsQuant function}
\usage{
gsQuant(sel, ebvs, sigma = 2, w = 0.5, nCrosses = 100,
  nProgenyPerCross = 1)
}
\arguments{
\item{sel}{[value]}

\item{ebvs}{[value]}

\item{sigma}{[value]. Default is 2}

\item{w}{[value]. Default is 0.5}

\item{nCrosses}{[value]. Default is 100}

\item{nProgenyPerCross}{[value]. Default is 1}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
