% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2toVe.R
\name{h2toVe}
\alias{h2toVe}
\title{h2toVe function}
\usage{
h2toVe(h2, Vg = 1)
}
\arguments{
\item{h2}{[value]}

\item{Vg}{[value]. Default is 1}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
