% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbreedFunc.R
\name{inbreedFunc}
\alias{inbreedFunc}
\title{inbreedFunc function}
\usage{
inbreedFunc(pop, GSfit, nProgPerFam, ...)
}
\arguments{
\item{pop}{[value]}

\item{GSfit}{[value]}

\item{nProgPerFam}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
