% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initFunc.R
\name{initFunc}
\alias{initFunc}
\title{initFunc function}
\usage{
initFunc(x)
}
\arguments{
\item{x}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
