% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertList.R
\name{invertList}
\alias{invertList}
\title{invertList function}
\usage{
invertList(ll)
}
\arguments{
\item{ll}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
