% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logSelInd.R
\name{logSelInd}
\alias{logSelInd}
\title{logSelInd function}
\usage{
logSelInd(pop, sel)
}
\arguments{
\item{pop}{[value]}

\item{sel}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
