% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxBv.R
\name{maxBv}
\alias{maxBv}
\title{maxBv function}
\usage{
maxBv(simParam, traits = 1)
}
\arguments{
\item{x}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
