% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxVar.R
\name{maxVar}
\alias{maxVar}
\title{maxVar function}
\usage{
maxVar(pop, GSfit, nSel, nCrosses, use, weightLoci = FALSE,
  pullGeno = pullSnpGeno, maxCrossPerParent = 0, verbose = FALSE,
  nProgeny = 1, ...)
}
\arguments{
\item{pop}{[value]}

\item{GSfit}{[value]}

\item{nSel}{[value]}

\item{nCrosses}{[value]}

\item{use}{[value]}

\item{weightLoci}{[value]. Default is FALSE}

\item{pullGeno}{[value]. Default is pullSnpGeno}

\item{maxCrossPerParent}{[value]. Default is 0}

\item{verbose}{[value]. Default is FALSE}

\item{nProgeny}{[value]. Default is 1}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
