% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePopsRec.R
\name{mergePopsRec}
\alias{mergePopsRec}
\title{mergePopsRec function}
\usage{
mergePopsRec(popList)
}
\arguments{
\item{x}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
