% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pherFuncID.R
\name{pherFuncID}
\alias{pherFuncID}
\title{pherFuncID function}
\usage{
pherFuncID(x, p)
}
\arguments{
\item{x}{[value]}

\item{p}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
