% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pherFuncMax.R
\name{pherFuncMax}
\alias{pherFuncMax}
\title{pherFuncMax function}
\usage{
pherFuncMax(x, p)
}
\arguments{
\item{x}{[value]}

\item{p}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
