% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPop.R
\name{plotPop}
\alias{plotPop}
\title{plotPop function}
\usage{
plotPop(simL, vLine = "none", popcol = "#000000", alpha = "80",
  alphaMean = "0D", ...)
}
\arguments{
\item{simL}{[value]}

\item{vLine}{[value]. Default is "none"}

\item{popcol}{[value]. Default is "#000000"}

\item{alpha}{[value]. Default is "0D"}

\item{alphaMean}{[value]. Default is "0D"}

\item{pch}{[value]. Default is 1}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
