% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popQuant.R
\name{popQuant}
\alias{popQuant}
\title{popQuant function}
\usage{
popQuant(sel, ebvs, sigma = 2, w = 0.5)
}
\arguments{
\item{sel}{[value]}

\item{ebvs}{[value]}

\item{sigma}{[value]. Default is 2}

\item{w}{[value]. Default is 0.5}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
