% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullLoci.R
\name{pullLoci}
\alias{pullLoci}
\title{pullLoci function}
\usage{
pullLoci(simParam, snpChip = 1, asList = FALSE)
}
\arguments{
\item{simParam}{[value]}

\item{snpChip}{[value]. Default is 1}

\item{asList}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
