% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullSegSites.R
\name{pullSegSites}
\alias{pullSegSites}
\title{pullSegSites function}
\usage{
pullSegSites(pop, returnMatrix = TRUE)
}
\arguments{
\item{pop}{[value]}

\item{returnMatrix}{[value]. Default is TRUE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
