% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{quant}
\alias{quant}
\title{quant function}
\usage{
quant(x, sigma, w)
}
\arguments{
\item{x}{[value]}

\item{sigma}{[value]}

\item{w}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
