% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomCross.R
\name{randomCross}
\alias{randomCross}
\title{randomCross function}
\usage{
randomCross(pop, nFam, nProgeny = 1)
}
\arguments{
\item{pop}{[value]}

\item{nFam}{[value]}

\item{nProgeny}{[value]. Default is 1}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
