% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlapply.R
\name{rlapply}
\alias{rlapply}
\title{rlapply function}
\usage{
rlapply(l, f = identity, level = 1, combine = list, counter = 1,
  ...)
}
\arguments{
\item{l}{[value]}

\item{f}{[value]. Default is identity}

\item{level}{[value]. Default is 1}

\item{combine}{[value]. Default is list}

\item{counter}{[value]. Default is 1}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
