% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSel.R
\name{rSel}
\alias{rSel}
\title{rSel function}
\usage{
rSel(sel)
}
\arguments{
\item{sel}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
