% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleFounderPop.R
\name{sampleFounderPop}
\alias{sampleFounderPop}
\title{sampleFounderPop function}
\usage{
sampleFounderPop(founderPop, size, n, savePop = FALSE, seed = NULL)
}
\arguments{
\item{founderPop}{[value]}

\item{size}{[value]}

\item{n}{[value]}

\item{savePop}{[value]. Default is FALSE}

\item{seed}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
