% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdUnCor.R
\name{sdUnCor}
\alias{sdUnCor}
\title{sdUnCor function}
\usage{
sdUnCor(x)
}
\arguments{
\item{x}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
