% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectInd2.R
\name{selectInd2}
\alias{selectInd2}
\title{selectInd2 function}
\usage{
selectInd2(pop, nSel, use, trait = 1, selFunc = identity)
}
\arguments{
\item{pop}{[value]}

\item{nSel}{[value]}

\item{use}{[value]}

\item{trait}{[value]. Default is 1}

\item{selFunc}{[value]. Default is identity}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
