% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDHdist.R
\name{simDHdist}
\alias{simDHdist}
\title{simDHdist function}
\usage{
simDHdist(nSel, pop, GSfit, retQuant = FALSE, quant = 0.9, nDH = 200,
  weight = 0.5, nProgeny = 1, returnPop = TRUE, bigmem = FALSE,
  ...)
}
\arguments{
\item{nSel}{[value]}

\item{pop}{[value]}

\item{GSfit}{[value]}

\item{retQuant}{[value]. Default is FALSE}

\item{quant}{[value]. Default is 0.9}

\item{nDH}{[value]. Default is 200}

\item{weight}{[value]. Default is 0.5}

\item{nProgeny}{[value]. Default is 1}

\item{returnPop}{[value]. Default is TRUE}

\item{bigmem}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
