% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDHdistPairs.R
\name{simDHdistPairs}
\alias{simDHdistPairs}
\title{simDHdistPairs function}
\usage{
simDHdistPairs(nSel, pop, GSfit, nCrosses, use, retQuant = FALSE,
  quant = 0.9, nDH = 200, weight = 0.5, maxCrossPerParent = 0,
  nSimCrosses = 1, nProgeny = 1, verbose = FALSE, ...)
}
\arguments{
\item{nSel}{[value]}

\item{pop}{[value]}

\item{GSfit}{[value]}

\item{nCrosses}{[value]}

\item{use}{[value]}

\item{retQuant}{[value]. Default is FALSE}

\item{quant}{[value]. Default is 0.9}

\item{nDH}{[value]. Default is 200}

\item{weight}{[value]. Default is 0.5}

\item{maxCrossPerParent}{[value]. Default is 0}

\item{nSimCrosses}{[value]. Default is 1}

\item{nProgeny}{[value]. Default is 1}

\item{verbose}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
