% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPlot.R
\name{simPlot}
\alias{simPlot}
\title{simPlot function}
\usage{
simPlot(popList, cols = "#000000", popLabs = NULL, varLine = "none",
  meanVariety = TRUE, legendPos = "topleft", plotReps = FALSE,
  plotVg = TRUE, plotSelInt = TRUE)
}
\arguments{
\item{popList}{[value]}

\item{cols}{[value]. Default is "#000000"}

\item{popLabs}{[value]. Default is NULL}

\item{varLine}{[value]. Default is "none"}

\item{meanVariety}{[value]. Default is TRUE}

\item{legendPos}{[value]. Default is "topleft"}

\item{plotReps}{[value]. Default is FALSE}

\item{plotVg}{[value]. Default is TRUE}

\item{plotSelInt}{[value]. Default is TRUE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
