% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSingleTraitInbred.R
\name{simSingleTraitInbred}
\alias{simSingleTraitInbred}
\title{simSingleTraitInbred function}
\usage{
simSingleTraitInbred(founderPop, paramL, simParam = SP,
  returnFunc = identity, verbose = TRUE, checkParam = FALSE,
  GSfunc = NULL, switchGSfunc = 4, ...)
}
\arguments{
\item{founderPop}{[value]}

\item{paramL}{[value]}

\item{simParam}{[value]. Default is SP}

\item{returnFunc}{[value]. Default is identity}

\item{verbose}{[value]. Default is TRUE}

\item{checkParam}{[value]. Default is FALSE}

\item{GSfunc}{[value]. Default is NULL}

\item{switchGSfunc}{[value]. Default is 4}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
