% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solqp.R
\name{solqp}
\alias{solqp}
\title{solqp function}
\usage{
solqp(pop, GSfit, use, nCrosses, simParam, lambda = NULL,
  fthresh = NULL, gain = NULL, truncqp = NULL, allowSelf = FALSE,
  weightLoci = FALSE, pullGeno = pullSnpGeno, verbose = FALSE,
  nProgeny = 1, maxProp = 1, ...)
}
\arguments{
\item{pop}{[value]}

\item{GSfit}{[value]}

\item{use}{[value]}

\item{nCrosses}{[value]}

\item{simParam}{[value]}

\item{lambda}{[value]. Default is NULL}

\item{fthresh}{[value]. Default is NULL}

\item{gain}{[value]. Default is NULL}

\item{truncqp}{[value]. Default is NULL}

\item{allowSelf}{[value]. Default is FALSE}

\item{weightLoci}{[value]. Default is FALSE}

\item{pullGeno}{[value]. Default is pullSnpGeno}

\item{verbose}{[value]. Default is FALSE}

\item{nProgeny}{[value]. Default is 1}

\item{maxProp}{[value]. Default is 1}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
