% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solqpOut.R
\name{solqpOut}
\alias{solqpOut}
\title{solqpOut function}
\usage{
solqpOut(pop, GSfit, use, nSel, nProgeny, nGenOut, nGenThisYr, simParam,
  limitN = 1, lambdaOut = NULL, fthreshOut = NULL, gainOut = NULL,
  truncqpOut = NULL, verbose = FALSE, ...)
}
\arguments{
\item{pop}{[value]}

\item{GSfit}{[value]}

\item{use}{[value]}

\item{nSel}{[value]}

\item{nProgeny}{[value]}

\item{nGenOut}{[value]}

\item{nGenThisYr}{[value]}

\item{simParam}{[value]}

\item{limitN}{[value]. Default is 0}

\item{lambdaOut}{[value]. Default is NULL}

\item{fthreshOut}{[value]. Default is NULL}

\item{gainOut}{[value]. Default is NULL}

\item{truncqpOut}{[value]. Default is NULL}

\item{verbose}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
