% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradSelCross.R
\name{tradSelCross}
\alias{tradSelCross}
\title{tradSelCross function}
\usage{
tradSelCross(pop, nCrosses, nFam, famSize, families, use, nProgeny = 1,
  intWithin = 0.2, intAcross = 1, equalWeight = FALSE,
  useFamPrior = FALSE, verbose = FALSE)
}
\arguments{
\item{pop}{[value]}

\item{nCrosses}{[value]}

\item{nFam}{[value]}

\item{famSize}{[value]}

\item{families}{[value]}

\item{use}{[value]}

\item{nProgeny}{[value]. Default is 1}

\item{intWithin}{[value]. Default is 0.2}

\item{intAcross}{[value]. Default is 1}

\item{equalWeight}{[value]. Default is FALSE}

\item{useFamPrior}{[value]. Default is FALSE}

\item{verbose}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
