% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradSelCross2.R
\name{tradSelCross2}
\alias{tradSelCross2}
\title{tradSelCross2 function}
\usage{
tradSelCross2(pop, elitepop, nCrosses, nFamilies, familySize, families,
  use, simParam = NULL, best = TRUE, nProgeny = 1, intWithin = 0.2,
  intAcross = 1, equalWeight = FALSE, useFamPrior = FALSE,
  verbose = FALSE)
}
\arguments{
\item{pop}{[value]}

\item{elitepop}{[value]}

\item{nCrosses}{[value]}

\item{nFamilies}{[value]}

\item{familySize}{[value]}

\item{families}{[value]}

\item{use}{[value]}

\item{best}{[value]. Default is TRUE}

\item{nProgeny}{[value]. Default is 1}

\item{intWithin}{[value]. Default is 0.2}

\item{intAcross}{[value]. Default is 1}

\item{equalWeight}{[value]. Default is FALSE}

\item{useFamPrior}{[value]. Default is FALSE}

\item{verbose}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
