% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncCross.R
\name{truncCross}
\alias{truncCross}
\title{truncCross function}
\usage{
truncCross(pop, nSel, nCrosses, use, nProgeny = 1,
  crossFunc = randomCross, traits = 1, ...)
}
\arguments{
\item{pop}{[value]}

\item{nSel}{[value]}

\item{nCrosses}{[value]}

\item{use}{[value]}

\item{nProgeny}{[value]. Default is 1}

\item{crossFunc}{[value]. Default is randomCross}

\item{traits}{[value]. Default is 1}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
