% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncSel.R
\name{truncSel}
\alias{truncSel}
\title{truncSel function}
\usage{
truncSel(pop, nSel, use, traits = 1, ...)
}
\arguments{
\item{pop}{[value]}

\item{nSel}{[value]}

\item{use}{[value]}

\item{traits}{[value]. Default is 1}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
