% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanRaden1.R
\name{vanRaden1}
\alias{vanRaden1}
\title{vanRaden1 function}
\usage{
vanRaden1(M)
}
\arguments{
\item{M}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
