% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedQuantile.R
\name{weightedQuantile}
\alias{weightedQuantile}
\title{weightedQuantile function}
\usage{
weightedQuantile(mu, sigmasq, quant = 0.9, w = 0.5)
}
\arguments{
\item{mu}{[value]}

\item{sigmasq}{[value]}

\item{quant}{[value]. Default is 0.9}

\item{w}{[value]. Default is 0.5}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
