% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withinFamSel.R
\name{withinFamSel}
\alias{withinFamSel}
\title{withinFamSel function}
\usage{
withinFamSel(pop, GSfit, use, int, nProgeny, nGenInbr, nGenThisYr,
  lambdaInbr = NULL, fthreshInbr = NULL, gainInbr = NULL,
  truncqpInbr = NULL, ssd, simParam, ...)
}
\arguments{
\item{pop}{[value]}

\item{GSfit}{[value]}

\item{use}{[value]}

\item{int}{[value]}

\item{nProgeny}{[value]}

\item{nGenInbr}{[value]}

\item{nGenThisYr}{[value]}

\item{lambdaInbr}{[value]. Default is NULL}

\item{fthreshInbr}{[value]. Default is NULL}

\item{gainInbr}{[value]. Default is NULL}

\item{truncqpInbr}{[value]. Default is NULL}

\item{ssd}{[value]}

\item{simParam}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
