% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrightsF.R
\name{wrightsF}
\alias{wrightsF}
\title{wrightsF function}
\usage{
wrightsF(M, returnNA = TRUE)
}
\arguments{
\item{M}{[value]}

\item{returnNA}{[value]. Default is TRUE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
